Added filtering to the backup crawler
The crawl method for the backup system now filters. It will find the
client directories, mark them, then find the correct year folder (if a
client has it) and mark that and any sub folders.
The result of this is that only the wanted files are included in the
tree and so only the wanted files are copied.

Also added override checking for files, so files are only copied if they
don't exist already or if override is enabled.